Remove circular dependencies in project form
