+++ src/pages/projects/form/ProjectFormPanel2.tsx
@@ -3,10 +3,9 @@ import { Field, FieldProps } from 'formik';
 import { useTranslation } from 'react-i18next';
 import { cristinCategories } from '../../../resources/cristinCategories';
 import { cristinKeywords } from '../../../resources/cristinKeywords';
-import { TypedLabel } from '../../../types/project.types';
+import { ProjectFieldName, TypedLabel } from '../../../types/project.types';
 import { dataTestId } from '../../../utils/dataTestIds';
 import { getLanguageString } from '../../../utils/translation-helpers';
-import { ProjectFieldName } from './ProjectFormDialog';
 import { RelatedProjectsField } from './RelatedProjectsField';
 
 export const ProjectFormPanel2 = () => {
